+++ generators/node/5-writing.js
@@ -3,12 +3,18 @@
 module.exports = function () {
 	// Create main files.
 	var mainFilePath = 'src/' + this.pkg.main || 'index.js';
-	this.fs.copy(
-		__dirname + '/templates/' + mainFilePath.replace('.js', '.benchmark.js'),
-		this.destinationPath(mainFilePath.replace('.js', '.benchmark.js'))
-	);
-	this.fs.copy(
-		__dirname + '/templates/' + mainFilePath.replace('.js', '.spec.js'),
-		this.destinationPath(mainFilePath.replace('.js', '.spec.js'))
-	);
+	
+	if (!this.fs.exists(mainFilePath)) {
+		this.fs.copy(
+			__dirname + '/templates/src/index.benchmark.js',
+			this.destinationPath(mainFilePath.replace('.js', '.benchmark.js'))
+		);
+	}
+	
+	if (!this.fs.exists(mainFilePath)) {
+		this.fs.copy(
+			__dirname + '/templates/src/index.spec.js',
+			this.destinationPath(mainFilePath.replace('.js', '.spec.js'))
+		);
+	}
 };
